feat(ivy): enable ng-reflect debug attributes (#27268)
PR Close #27268
