+++ packages/platform-server/test/integration_spec.ts
@@ -614,15 +614,14 @@ class HiddenModule {
            });
          }));
 
-      fixmeIvy('to investigate') &&
-          it('should handle false values on attributes', async(() => {
-               renderModule(FalseAttributesModule, {document: doc}).then(output => {
-                 expect(output).toBe(
-                     '<html><head></head><body><app ng-version="0.0.0-PLACEHOLDER">' +
-                     '<my-child ng-reflect-attr="false">Works!</my-child></app></body></html>');
-                 called = true;
-               });
-             }));
+      it('should handle false values on attributes', async(() => {
+           renderModule(FalseAttributesModule, {document: doc}).then(output => {
+             expect(output).toBe(
+                 '<html><head></head><body><app ng-version="0.0.0-PLACEHOLDER">' +
+                 '<my-child ng-reflect-attr="false">Works!</my-child></app></body></html>');
+             called = true;
+           });
+         }));
 
       it('should handle element property "name"', async(() => {
            renderModule(NameModule, {document: doc}).then(output => {
